Completed Part 3 of the Hangman application : "The Game With HInts"
The user is now allowed to guess the special character "*" to get a hint of all the possible words the secret word can be based on the letters they have already guessed. For example, if the secret word is "apple" and the user has already guessed "a _ _ l e" but would like a hint, the program will output all the words that can possibly match that one (i.e. agile, ample, ...) given the letters already guessed and letters not yet guessed.