Added 'Problem Set 3' to repository
'Problem Set 3' is a problem with a word game, similar to Scrabble or Words With Friends. Letters are dealt to players, who then construct one or more words using their letters. Each valid word earns the user points, based on the length of the word and the
letters in that word.